requirements.txt
# WhatsApp API Microservice - Dependencias

# Framework principal
Flask==3.0.0
Flask-RESTX==1.3.0

# Base de datos
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# PostgreSQL (para producción)
psycopg2-binary==2.9.9

# Redis para cache y rate limiting
redis==5.0.1

# CORS y seguridad
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# HTTP requests para WhatsApp API
requests==2.31.0
urllib3==2.1.0

# Utilidades
python-dotenv==1.0.0
click==8.1.7

# Validación de datos
marshmallow==3.20.1
pydantic==2.5.0

# Logging y monitoreo
structlog==23.2.0
colorlog==6.8.0  # Para logs coloridos en terminal

# Desarrollo y testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Seguridad
cryptography==41.0.8
PyJWT==2.8.0

# Archivos y media
Pillow==10.1.0

# Chatbot y AI
rivescript==1.15.0
openai==1.3.5  # Para LLM opcional

# Formato y linting (desarrollo)
black==23.11.0
flake8==6.1.0
isort==5.12.0
